+++ src/bookshelf/bookshelf.spec.ts
@@ -417,6 +417,47 @@ describe('Note processing', () => {
             '2025-01-01: Dracula: started',
         ])
     })
+
+    test('Moving a book note to a different folder should remove the corresponding book', async () => {
+        const shining = new FakeNote('Books/The Shining.md', new StaticMetadata({}), [])
+        const animalFarm = new FakeNote('Books/Animal Farm.md', new StaticMetadata({}), [])
+        const dracula = new FakeNote('Books/Dracula.md', new StaticMetadata({}), [])
+        await bookshelf.process(shining)
+        await bookshelf.process(animalFarm)
+        await bookshelf.process(dracula)
+
+        animalFarm.path = 'Notes/Animal Farm.md'
+        await bookshelf.process(animalFarm)
+
+        expect(Array.from(bookshelf.all()).map((b) => b.metadata.title)).toEqual(['The Shining', 'Dracula'])
+    })
+
+    test('Moving a book note to a different folder should remove the related reading process', async () => {
+        const shining = new FakeNote('Books/The Shining.md', new StaticMetadata({}), [])
+        const animalFarm = new FakeNote('Books/Animal Farm.md', new StaticMetadata({}), [])
+        const dracula = new FakeNote('Books/Dracula.md', new StaticMetadata({}), [])
+        notes.add(shining)
+        notes.add(animalFarm)
+        notes.add(dracula)
+        await bookshelf.process(shining)
+        await bookshelf.process(animalFarm)
+        await bookshelf.process(dracula)
+        await bookshelf.process(
+            new FakeNote('2025-01-01.md', new StaticMetadata({}), [
+                'Started reading [[The Shining]]',
+                'Started reading [[Animal Farm]]',
+                'Started reading [[Dracula]]',
+            ]),
+        )
+
+        animalFarm.path = 'Notes/Animal Farm.md'
+        await bookshelf.process(animalFarm)
+
+        expect(bookshelf.readingJourney().map(readingProgressAsString)).toEqual([
+            '2025-01-01: The Shining: started',
+            '2025-01-01: Dracula: started',
+        ])
+    })
 })
 
 test('It should return all books added to the bookshelf', async () => {